implement top_ten_temps method
implemented the top_ten_temps method . an SQL query is used to query the database and returns the date and times the hottest temperatures occurred in descending order of outside temperatures. The method iterates through first 10 items as this is the expected return. Currently the list is not sorted by date.