Implement hi_low_temps
implemented hi_low_temps method. An SQL query is used to query Hi Temperature and return if the value is +- 1 degree of 22.3 or the Low Temperature value is +- 0.2 degrees of 10.3. The low temperature results are iterated and appended to a list if the date is any of the first 9 days of June.